% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCplot.R
\name{MCplot}
\alias{MCplot}
\title{Plot of power function.}
\usage{
MCplot(
  object,
  nstats = c("W", "LR", "RS", "G", "abs", "sqs", "St"),
  alpha = 0.05,
  xlim = 5,
  legend_title = "Test",
  tag_title = ggplot2::element_blank(),
  name_title = ggplot2::element_blank(),
  tcolor = FALSE
)
}
\arguments{
\item{object}{An object of class "\code{MCplot}"}

\item{nstats}{Name(s) of test statistics to be plotted. Note a maximum of 4 tests allowed. By default c("W", "LR", "RS", "G") will be plotted.}

\item{alpha}{Probability of error of first kind (in plot shown as dotted horizontal line).}

\item{xlim}{A numeric value, specifying the left/lower limit and the right/upper limit of the x scale. Maximum value is xlim = 10.}

\item{legend_title}{The text for Legend. By default it is set to "Test".}

\item{tag_title}{The text for tag label, which will be displayed at the top-left of the plot, by default of Item 1 only. By default it is element_blank().}

\item{name_title}{The text for title of plot. By default it is element_blank().}

\item{tcolor}{Logic value. If TRUE then ggplot plot will be in color otherwise black and white.}
}
\value{
An object of class "\code{MCplot}" containing:
 \item{plotlist}{A list of length k containing ggplot objects of power plots of each item and selected test statistics.}
 \item{call}{The matched call.}
}
\description{
This function plots the power function of selected test statistics
}
\examples{
\dontrun{
# Rasch model with sample size N = 60, k = 8
library(tclboot)

# model parameters
items <- c(0, -2, -1, -0.5, 0, 0.5, 1, 2)
x <- c(rep(0,25), rep(1,35))

# generate initial data matrix
y <- MCsimRasch(N = 60, splitcr = x, items = items)$X

obj <- MCboot(X=y, splitcr = x, n_iter = 1000)

res <- MCplot(object = obj, nstats = c("W", "G"))

}
}
\references{
{
Draxler, C., & Dahm, S. (2020). Conditional or Pseudo Exact Tests with an Application in the Context of Modeling
Response Times. Psych, 2(4), 198-208. \url{https://doi.org/10.3390/psych2040017}
}
}
\keyword{mcmc}
