% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCplot_simdata_crit.R
\name{MCplot_simdata_crit}
\alias{MCplot_simdata_crit}
\title{Plot of power function.}
\usage{
MCplot_simdata_crit(
  N,
  k,
  nstats = c("W", "LR", "RS", "G", "U1", "U2", "St"),
  alpha = 0.05,
  xlim = 5,
  legend_title = ggplot2::element_blank(),
  tag_title = ggplot2::element_blank(),
  name_title = ggplot2::element_blank(),
  legend.position = c(0.5, 0.75),
  legend.direction = "vertical",
  tcolor = FALSE,
  ctr_alpha = 1,
  ctr_mod = c("low", "high", "both")
)
}
\arguments{
\item{N}{Sample size \eqn{n} of selected case.  N =  10, 15, 20, 25, 30, 50}

\item{k}{Number of items \eqn{k} of selected case. k = 6, 8, 10, 12, 14, 16, 18, 20, 25, 30}

\item{nstats}{Name of test statistics chosen out of possible tests c("W", "LR", "RS", "G", "U1", "U2", "St") to be modified. Note only one test allowed. By default "U1" will be plotted.}

\item{alpha}{Probability of error of first kind (in plot shown as dotted horizontal line).}

\item{xlim}{A numeric value, specifying the left/lower limit and the right/upper limit of the x scale. Maximum value is xlim = 10.}

\item{legend_title}{The text for Legend. By default it is set to "Test".}

\item{tag_title}{The text for tag label, which will be displayed at the top-left of the plot, by default of Item 1 only. By default it is element_blank().}

\item{name_title}{The text for title of plot. By default it is element_blank().}

\item{legend.position}{the position of legends ("none", "left", "right", "bottom", "top", or two-element numeric vector)}

\item{legend.direction}{layout of items in legends ("horizontal" or "vertical")}

\item{tcolor}{Logic value. If TRUE then ggplot plot will be in color otherwise black and white.}

\item{ctr_alpha}{A numeric value. The crit. region \eqn{C} of selected test statistic will be \eqn{100(alpha*ctr_alpha)} percent largest values of this test statistics. Default value is 1.}

\item{ctr_mod}{Character. Which type of modification of critical region should be used. Adding low most extreme values of the respective of \eqn{T} statistics ("low"),
adding high most extreme values of the respective of \eqn{T} statistics ("high"), or adding both ("both"). The latter is also Default value.}
}
\value{
An object of class "\code{MCplot_case_crit}" containing:
\item{power_item}{A list of length k containing ggplot objects of power plots of each item and selected test statistics and
 modification of critical region (cond).}
\item{hist_item}{Histograms of critical values of sufficient statistics \eqn{t}, without ("test name") and with (mod) modification of critical region.}
 \item{plotlist}{A list of length k containing combined ggplot objects of (1) power plots of each item and selected test statistics and
 modification of critical region, and (2) Histograms of critical values of sufficient statistics \eqn{t}, without ("test name") and with (mod) modification
 of critical region.}
 \item{resultlist}{A list of length k containing several sublists: values of sufficient statistics \eqn{t} "t_item",
 indices of \eqn{100(alpha*ctr_alpha)} percent largest values ("icrit"), and \eqn{100(alpha)} percent largest values of test statistic ("icrit0"),
 incices of modification ("imod", "imod_low", "imod_high"), difference of indices ("idiff"), union of indices in crit. reg. after modification ("icrit_mod"),
 power values at \eqn{delta}=0 ("lpwr_d0"), and a table of rel. frequency of values of sufficient statistics \eqn{t} in critical region ("t_table") }
 \item{hist_crit_reg}{A list of length k containing histograms of critical values of sufficient statistics \eqn{t}, without modification ("test name") and with (mod) modification of critical region.}
  \item{call}{The matched call.}
}
\description{
This function plots the power function of selected test statistics
}
\examples{
\dontrun{

# load MC sim data of case G-V-30x18 (with N=30, k=18)
#  and generate power plots of selected test and
#  with modification of critical region

res <- MCplot_simdata_crit(N=30, k=18, nstats = "G")


}
}
\references{
{
Draxler, C., & Dahm, S. (2020). Conditional or Pseudo Exact Tests with an Application in the Context of Modeling
Response Times. Psych, 2(4), 198-208. \url{https://doi.org/10.3390/psych2040017}
}
}
\keyword{mcmc}
