% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_plot.R
\name{save_plot}
\alias{save_plot}
\title{Utility function for saving a plotlist of ggplot objects}
\usage{
save_plot(plotlist, sub_dir = "plot", scale = 1, px = 80, py = 80)
}
\arguments{
\item{plotlist}{A list of length k containing ggplot objects of power plots.}

\item{sub_dir}{Sub directory to save plot to. \code{sub_dir} is added by default to current working directory.}

\item{scale}{Multiplicative scaling factor used in \code{ggplot}. Default = 1.}

\item{px}{Plot size in "mm". Default = 80 mm.}

\item{py}{Plot size in "mm". Default = 80 mm.}
}
\value{
A list of results.
\item{call}{The matched call.}
}
\description{
This utility function saves individual plots of a plotlist of ggplot objects
 using \pkg{ggplot2} function \code{ggsave}.
}
\examples{
\dontrun{
#

res <- MCplot_simdata_ext(npers = c(10,20, 30, 50), k=14, nstats = "sqs", xlim = 4, tcolor = TRUE)

res <- save_plot(plotlist = res$plotlist, sub_dir = "plot_k=14")

}
}
\references{
{
none
}
}
\seealso{
\code{\link{MCboot}}, \code{\link{MCplot}} and \code{\link{MCplot_crit}}
}
\keyword{mcmc}
